# [1.19] Improve duplicate checks for implicit counts of 1