--- a/Common/src/main/java/com/almostreliable/unified/recipe/RecipeLink.java
+++ b/Common/src/main/java/com/almostreliable/unified/recipe/RecipeLink.java
@@ -46,8 +46,8 @@ public class RecipeLink {
 
         JsonObject compare = null;
         if (first.getType().toString().equals("minecraft:crafting_shaped")) {
-            compare = JsonCompare.compareShaped(selfActual, toCompareActual, compareSettings.getIgnoredFields());
-        } else if (JsonCompare.matches(selfActual, toCompareActual, compareSettings.getIgnoredFields())) {
+            compare = JsonCompare.compareShaped(selfActual, toCompareActual, compareSettings);
+        } else if (JsonCompare.matches(selfActual, toCompareActual, compareSettings)) {
             compare = JsonCompare.compare(compareSettings.getRules(), selfActual, toCompareActual);
         }