--- a/Common/src/test/java/com/almostreliable/unified/util/JsonCompareTest.java
+++ b/Common/src/test/java/com/almostreliable/unified/util/JsonCompareTest.java
@@ -12,7 +12,7 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.*;
 
 public class JsonCompareTest {
-    public static String recipe = """
+    private static final String RECIPE = """
             {
               "type": "minecraft:smelting",
               "group": "coal",
@@ -27,8 +27,8 @@ public class JsonCompareTest {
 
     @Test
     public void simpleCompareFirst() {
-        JsonObject first = TestUtils.json(recipe, j -> j.addProperty("experience", 0.2));
-        JsonObject second = TestUtils.json(recipe); // 0.1 experience
+        JsonObject first = TestUtils.json(RECIPE, j -> j.addProperty("experience", 0.2));
+        JsonObject second = TestUtils.json(RECIPE); // 0.1 experience
 
         LinkedHashMap<String, JsonCompare.Rule> rules = new LinkedHashMap<>();
         rules.put("experience", new JsonCompare.LowerRule());
@@ -38,8 +38,8 @@ public class JsonCompareTest {
 
     @Test
     public void simpleCompareSecond() {
-        JsonObject first = TestUtils.json(recipe, j -> j.addProperty("experience", 0.05));
-        JsonObject second = TestUtils.json(recipe); // 0.1 experience
+        JsonObject first = TestUtils.json(RECIPE, j -> j.addProperty("experience", 0.05));
+        JsonObject second = TestUtils.json(RECIPE); // 0.1 experience
 
         LinkedHashMap<String, JsonCompare.Rule> rules = new LinkedHashMap<>();
         rules.put("experience", new JsonCompare.LowerRule());
@@ -49,8 +49,8 @@ public class JsonCompareTest {
 
     @Test
     public void compareHigherWins() {
-        JsonObject first = TestUtils.json(recipe, j -> j.addProperty("experience", 0.05));
-        JsonObject second = TestUtils.json(recipe); // 0.1 experience  // 0.1 experience
+        JsonObject first = TestUtils.json(RECIPE, j -> j.addProperty("experience", 0.05));
+        JsonObject second = TestUtils.json(RECIPE); // 0.1 experience  // 0.1 experience
 
         LinkedHashMap<String, JsonCompare.Rule> rules = new LinkedHashMap<>();
         rules.put("experience", new JsonCompare.HigherRule());
@@ -60,19 +60,19 @@ public class JsonCompareTest {
 
     @Test
     public void compareMulti() {
-        JsonObject a = TestUtils.json(recipe, j -> {
+        JsonObject a = TestUtils.json(RECIPE, j -> {
             j.addProperty("experience", 0.1);
             j.addProperty("cookingtime", 100);
         });
-        JsonObject b = TestUtils.json(recipe, j -> j.addProperty("experience", 0.1));
-        JsonObject c = TestUtils.json(recipe, j -> {
+        JsonObject b = TestUtils.json(RECIPE, j -> j.addProperty("experience", 0.1));
+        JsonObject c = TestUtils.json(RECIPE, j -> {
             j.addProperty("experience", 0.1);
             j.addProperty("cookingtime", 50);
         });
-        JsonObject d = TestUtils.json(recipe, j -> j.addProperty("experience", 0.2));
-        JsonObject e = TestUtils.json(recipe, j -> j.addProperty("experience", 0.2));
-        JsonObject f = TestUtils.json(recipe, j -> j.addProperty("experience", 0.1));
-        JsonObject g = TestUtils.json(recipe, j -> {
+        JsonObject d = TestUtils.json(RECIPE, j -> j.addProperty("experience", 0.2));
+        JsonObject e = TestUtils.json(RECIPE, j -> j.addProperty("experience", 0.2));
+        JsonObject f = TestUtils.json(RECIPE, j -> j.addProperty("experience", 0.1));
+        JsonObject g = TestUtils.json(RECIPE, j -> {
             j.addProperty("experience", 0.2);
             j.addProperty("cookingtime", 100);
         });
@@ -91,24 +91,24 @@ public class JsonCompareTest {
 
     @Test
     public void simpleMatch() {
-        JsonObject first = TestUtils.json(recipe);
-        JsonObject second = TestUtils.json(recipe);
+        JsonObject first = TestUtils.json(RECIPE);
+        JsonObject second = TestUtils.json(RECIPE);
         boolean matches = JsonCompare.matches(first, second, List.of());
         assertTrue(matches);
     }
 
     @Test
     public void noMatch() {
-        JsonObject first = TestUtils.json(recipe, j -> j.addProperty("experience", 100));
-        JsonObject second = TestUtils.json(recipe);
+        JsonObject first = TestUtils.json(RECIPE, j -> j.addProperty("experience", 100));
+        JsonObject second = TestUtils.json(RECIPE);
         boolean matches = JsonCompare.matches(first, second, List.of());
         assertFalse(matches);
     }
 
     @Test
     public void matchBecauseIgnore() {
-        JsonObject first = TestUtils.json(recipe, j -> j.addProperty("experience", 100));
-        JsonObject second = TestUtils.json(recipe);
+        JsonObject first = TestUtils.json(RECIPE, j -> j.addProperty("experience", 100));
+        JsonObject second = TestUtils.json(RECIPE);
         boolean matches = JsonCompare.matches(first, second, List.of("experience"));
         assertTrue(matches);
     }
@@ -202,5 +202,80 @@ public class JsonCompareTest {
         JsonObject second = TestUtils.json(recipe2);
         JsonObject result = JsonCompare.compareShaped(first, second, List.of("pattern", "key"));
         assertEquals(first, result);
+    }
+
+    @Test
+    public void sanitizeImplicitCount() {
+        String recipe1 = """
+                {
+                  "type": "minecraft:crafting_shaped",
+                  "pattern": [
+                    "iii",
+                    "iii",
+                    "iii"
+                  ],
+                  "key": {
+                    "i": {
+                      "tag": "forge:raw_materials/iron"
+                    }
+                  },
+                  "result": "minecraft:iron_ingot"
+                }
+                """;
+        String recipe2 = """
+                {
+                  "type": "minecraft:crafting_shaped",
+                  "pattern": [
+                    "iii",
+                    "iii",
+                    "iii"
+                  ],
+                  "key": {
+                    "i": {
+                      "tag": "forge:raw_materials/iron"
+                    }
+                  },
+                  "result": {
+                    "item": "minecraft:iron_ingot",
+                    "count": 1
+                  }
+                }
+                """;
+
+        JsonObject first = TestUtils.json(recipe1);
+        JsonObject second = TestUtils.json(recipe2);
+        JsonObject result = JsonCompare.compareShaped(first, second, List.of("pattern", "key"));
+        assertEquals(first, result);
+    }
+
+    @Test
+    public void sanitizeImplicitCountNested() {
+        String recipe1 = """
+                {
+                  "type": "create:crushing",
+                  "ingredients": [{ "tag": "forge:raw_materials/lead" }],
+                  "processingTime": 400,
+                  "results": [
+                    { "item": "emendatusenigmatica:crushed_lead_ore" },
+                    { "chance": 0.75, "item": "create:experience_nugget" }
+                  ]
+                }
+                """;
+        String recipe2 = """
+                {
+                  "type": "create:crushing",
+                  "ingredients": [{ "tag": "forge:raw_materials/lead" }],
+                  "processingTime": 400,
+                  "results": [
+                    { "count": 1, "item": "emendatusenigmatica:crushed_lead_ore" },
+                    { "chance": 0.75, "count": 1, "item": "create:experience_nugget" }
+                  ]
+                }
+                """;
+
+        JsonObject first = TestUtils.json(recipe1);
+        JsonObject second = TestUtils.json(recipe2);
+        boolean result = JsonCompare.matches(first, second, List.of());
+        assertTrue(result);
     }
 }